add project handler
