+++ src/main/java/com/edj/teamoop/exception/handler/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
-package com.edj.teamoop.exception;
+package com.edj.teamoop.exception.handler;
 
+import com.edj.teamoop.exception.ProjectNotFoundException;
 import com.edj.teamoop.response.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -30,14 +31,5 @@ public class GlobalExceptionHandler {
         return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
     }
 
-    @ExceptionHandler(ProjectNotFoundException.class)
-    public ResponseEntity<ErrorResponse> handleProjectNotFoundException(ProjectNotFoundException ex, WebRequest request) {
-        ErrorResponse errorResponse = new ErrorResponse(
-                ex.getMessage(),
-                request.getDescription(false),
-                HttpStatus.NOT_FOUND.value()
-        );
-        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
-    }
 }
